Change project package name.
