+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "29.0.2"
 
     defaultConfig {
-        applicationId "io.github.entimer.coronatracker"
+        applicationId "com.entimer.coronatracker"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 10
