+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.github.entimer.coronatracker">
+    package="com.entimer.coronatracker">
 
     <uses-permission android:name="android.permission.INTERNET"/>
 
@@ -11,15 +11,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.splash.SplashActivity">
+        <activity android:name="com.entimer.coronatracker.ui.splash.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.main.MainActivity"/>
-        <activity android:name=".ui.setting.SettingActivity"/>
+        <activity android:name="com.entimer.coronatracker.ui.main.MainActivity"/>
+        <activity android:name="com.entimer.coronatracker.ui.setting.SettingActivity"/>
     </application>
 
 </manifest>
