+++ app/src/main/java/com/entimer/coronatracker/api/country/Country.kt
@@ -1,0 +1,3 @@
+package com.entimer.coronatracker.api.country
+
+data class Country(val name: String, val code: String)
