+++ app/src/main/java/com/entimer/coronatracker/api/covid/CaseData.kt
@@ -1,0 +1,8 @@
+package com.entimer.coronatracker.api.covid
+
+data class CaseData(
+    val date: String,
+    val confirmed: Int,
+    val recovered: Int,
+    val death: Int
+)
