+++ app/src/main/java/com/entimer/coronatracker/api/iso3166/Iso3166Data.kt
@@ -1,0 +1,8 @@
+package com.entimer.coronatracker.api.iso3166
+
+data class Iso3166Data(
+    val name: String,
+    val alpha2: String,
+    val alpha3: String,
+    val numeric: String
+)
