+++ app/src/main/java/com/entimer/coronatracker/ui/base/IView.kt
@@ -1,0 +1,6 @@
+package com.entimer.coronatracker.ui.base
+
+interface IView {
+    fun initViews()
+    fun initListeners()
+}
