+++ app/src/main/java/com/entimer/coronatracker/ui/main/map/MapFragment.kt
@@ -1,0 +1,14 @@
+package com.entimer.coronatracker.ui.main.map
+
+import android.app.Fragment
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.entimer.coronatracker.R
+
+class MapFragment: Fragment() {
+    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater!!.inflate(R.layout.fragment_map, container, false)
+    }
+}
