--- app/src/main/java/io/github/entimer/coronatracker/api/country/Country.kt
@@ -1,3 +1,0 @@
-package io.github.entimer.coronatracker.api.country
-
-data class Country(val name: String, val code: String)
