--- app/src/main/java/io/github/entimer/coronatracker/api/covid/CaseData.kt
@@ -1,8 +1,0 @@
-package io.github.entimer.coronatracker.api.covid
-
-data class CaseData(
-    val date: String,
-    val confirmed: Int,
-    val recovered: Int,
-    val death: Int
-)
