--- app/src/main/java/io/github/entimer/coronatracker/api/iso3166/Iso3166Data.kt
@@ -1,8 +1,0 @@
-package io.github.entimer.coronatracker.api.iso3166
-
-data class Iso3166Data(
-    val name: String,
-    val alpha2: String,
-    val alpha3: String,
-    val numeric: String
-)
