--- app/src/main/java/io/github/entimer/coronatracker/room/dao/Iso3166Dao.kt
@@ -1,28 +1,0 @@
-package io.github.entimer.coronatracker.room.dao
-
-import androidx.room.*
-import io.github.entimer.coronatracker.room.entity.Iso3166Entity
-
-@Dao
-interface Iso3166Dao {
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun insert(iso3166Entity: Iso3166Entity)
-
-    @Update
-    fun update(iso3166Entity: Iso3166Entity)
-
-    @Delete
-    fun delete(iso3166Entity: Iso3166Entity)
-
-    @Query("SELECT * FROM iso3166")
-    fun selectAll(): List<Iso3166Entity>
-
-    @Query("SELECT * FROM iso3166 WHERE name = (:name)")
-    fun selectByName(name: String): Iso3166Entity
-
-    @Query("SELECT * FROM iso3166 WHERE alpha3 = (:alpha3)")
-    fun selectByAlpha3(alpha3: String): Iso3166Entity
-
-    @Query("SELECT * FROM iso3166 WHERE name IN (:keyword) OR alpha3 IN (:keyword)")
-    fun selectByKeyword(keyword: String): List<Iso3166Entity>
-}
