--- app/src/main/java/io/github/entimer/coronatracker/ui/base/IView.kt
@@ -1,6 +1,0 @@
-package io.github.entimer.coronatracker.ui.base
-
-interface IView {
-    fun initViews()
-    fun initListeners()
-}
