--- app/src/main/java/io/github/entimer/coronatracker/ui/setting/SettingActivity.kt
@@ -1,17 +1,0 @@
-package io.github.entimer.coronatracker.ui.setting
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import io.github.entimer.coronatracker.R
-import kotlinx.android.synthetic.main.activity_setting.*
-
-class SettingActivity: AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_setting)
-
-        setting_toolbar.setNavigationOnClickListener {
-            finish()
-        }
-    }
-}
