+++ app/src/main/res/layout/activity_setting.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.setting.SettingActivity">
+    tools:context="com.entimer.coronatracker.ui.setting.SettingActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/setting_toolbar"
